Добавлены исключения в сервисы.

+++ src/main/java/com/skypro/animalShelterInfoBot/services/AnimalService.java
@@ -17,12 +17,18 @@ import java.util.List;
             this.animalRepository = animalRepository;
         }
         public Animal createAnimal(Animal animal) {
+            if (animal == null) {
+                throw new IllegalArgumentException("Animal cannot be null");
+            }
             return animalRepository.save(animal);
         }
     public Animal updateAnimal(long id, Animal animal) {
+        if (animal == null) {
+            throw new IllegalArgumentException("Animal cannot be null");
+        }
         Animal updatedAnimal = animalRepository.findAnimalById(id);
         if (updatedAnimal == null) {
-            return null;
+            throw new IllegalArgumentException("Animal with id " + id + " not found");
         }
         updatedAnimal.setCatOrDog(animal.getCatOrDog());
         updatedAnimal.setNickName(animal.getNickName());
@@ -38,7 +44,7 @@ import java.util.List;
 
                 return animals.size();
             } else {
-                return animals;
+                throw new IllegalArgumentException("Page number and size number must not be null");
             }
         }
 
@@ -48,6 +54,9 @@ import java.util.List;
             return animalRepository.findAll();
         }
         public void deleteAnimalById(Long Id) {
+            if (Id == null) {
+                throw new IllegalArgumentException("Id cannot be null");
+            }
             animalRepository.deleteById(Id);
         }
 }
